// src/app/sylabus-page/syllabus.interface.ts
// syllabus.interface.ts
export interface Syllabus {
  literature: string;
  teacherInfo: string;
  prerequisites: string;
  corequisites: string;
  disciplineGoal: string;
  disciplineContent: string;
  individualTasks: string;
  software: string;
  studyResults: string;
  additionalInfo: string;
}